models/X-VLM/requirements.txt
timm==0.4.9
transformers==4.12.5
ruamel_yaml
opencv-python
scikit-image
matplotlib
torch==1.10.0
torchvision==0.11.1
pycocotools
pycocoevalcap
